Website Updates
Edited comments in the services and index html file. Updated names of the services and the index page to jtc_services.html and jtc_index.html respectively. Removed (1) from file names to match. Added links to jtc_base.css and jtc_layout.css in the index.html. The link and meta tags were added to the head of the index document. Aside, Header, and Article elements were added to the index body. A navigation bar was added in the header element. Quotes were added in the aside using block quotes. Text was added in the article in order to add text. Information at the bottom of the index page was added in an address element in the index. The email and telephone number in that address element were set to links. Meta and link information was added to the services page. The index header was copied over to the services page so they have the same header. The text was added underneath the header tag on the services page. In that text the title was made a h1. An image was put after that h1. The rest of the text was made into paragraph, and the subtitle that went with that text was made h2. A Contact footer was added to the end of the services page.